+++ sei-db/state_db/sc/migration/router_kvstore_test.go
@@ -215,7 +215,7 @@ func (f *failingRouter) GetProof(string, []byte) (*ics23.CommitmentProof, error)
 	return nil, errors.New("failingRouter.GetProof: not used by these tests")
 }
 
-func (f *failingRouter) SetMigrationBatchSize(uint64) {}
+func (f *failingRouter) SetMigrationBatchSize(int) {}
 
 func TestRouterCommitKVStore_GetPanicsOnRouterError(t *testing.T) {
 	store := NewRouterCommitKVStore(
